promo_service/requirements.txt
Django==4.1.7
django-admin-rangefilter==0.9.0
django-import-export==3.1.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
gunicorn==20.1.0
python-dotenv==1.0.0
python-gettext==4.1
